+++ src/driver/input/GT911.cpp
@@ -26,10 +26,11 @@ namespace be::driver::input {
     } ;
 
     GT911::GT911(JSContext * ctx, be::I2C * i2c, uint8_t addr)
-        : InDevPointer(ctx,build(ctx))
+        : InDevPointer(ctx,build(ctx),5)
         , addr(addr)
         , i2c(i2c)
-    {}
+    {
+    }
 
     JSValue GT911::constructor(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
         ARGV_TO_UINT8_OPT(0, busnum, 255)
@@ -74,7 +75,7 @@ namespace be::driver::input {
         return data & 0x80 ;
     }
 
-    bool GT911::readPos(uint8_t i, uint16_t &x, uint16_t &y) {
+    bool GT911::readPos(uint8_t i, int16_t &x, int16_t &y) {
         if(!i2c) {
             return false ;
         }
@@ -150,7 +151,14 @@ namespace be::driver::input {
             JSTHROW("invalid GT911 address:0x%02X (must be 0x5D or 0x14)", thisobj->addr)
         }
 
+        // @todo read from setup options
+        thisobj->pointsSize = 1 ;
+        if(thisobj->pointsSize>INDEVPOINTER_MAX_NUM) {
+            thisobj->pointsSize = INDEVPOINTER_MAX_NUM ;
+        }
+
         thisobj->reset() ;
-        return JS_UNDEFINED ;
+
+        return InDevPointer::setup(ctx, this_val, argc, argv) ;
     }
 }