+++ src/ext/media/audio/stream/el_src.c
@@ -1,15 +1,23 @@
 #include "audio_stream.h"
+#include <string.h>
 
 
 audio_el_src_t *  audio_el_src_create(audio_pipe_t * pipe, uint8_t core) {
     audio_el_src_t * el ;
-    ELEMENT_CREATE(pipe, audio_el_src_t, el, task_src, 1024*3, 5, core, 1024*2)
+    ELEMENT_CREATE(pipe, audio_el_src_t, el, task_src, 1024*3, 10, core, 1024*2)
     el->base.name = "src" ;
     return el ;
 }
 
 bool audio_el_src_strip_mp3(audio_el_src_t * el) {
-    if(!audio_el_src_strip_pcm(el)){
+
+    if(el->file){
+        fclose(el->file) ;
+        el->file = NULL ;
+    }
+    el->file = fopen(el->src_path,"rb") ;
+    if(!el->file) {
+        printf("can not open file: %s", el->src_path) ;
         return false ;
     }
 
@@ -34,6 +42,20 @@ bool audio_el_src_strip_mp3(audio_el_src_t * el) {
 }
 
 
+struct WavHeader {
+    char riffHeader[4];    // "RIFF"
+    uint32_t fileSize;     // 文件大小
+    char waveHeader[4];    // "WAVE"
+    char fmtHeader[4];     // "fmt "
+    uint32_t fmtChunkSize; // 格式块大小（通常为 16）
+    uint16_t audioFormat;  // 音频格式（1 = PCM）
+    uint16_t numChannels;  // 通道数
+    uint32_t sampleRate;   // 采样率
+    uint32_t byteRate;     // 每秒字节数
+    uint16_t blockAlign;   // 块对齐
+    uint16_t bitsPerSample;// 每个样本的位数
+};
+
 bool audio_el_src_strip_pcm(audio_el_src_t * el) {
 
     if(el->file){
@@ -46,7 +68,55 @@ bool audio_el_src_strip_pcm(audio_el_src_t * el) {
         return false ;
     }
 
-    return true ;
+    struct WavHeader header ;
+
+    // 读取 WAV 文件头
+    if (fread(&header, sizeof(struct WavHeader), 1, el->file) != 1) {
+        printf("read wav header faild.") ;
+        fclose(el->file);
+        return false;
+    }
+
+    // 检查 RIFF 和 WAVE 标识是否合法
+    if (strncmp(header.riffHeader, "RIFF", 4) != 0 || strncmp(header.waveHeader, "WAVE", 4) != 0) {
+        printf("invalid wav file.") ;
+        fclose(el->file);
+        el->file = NULL ;
+        return false ;
+    }
+
+    i2s_set_clk((i2s_port_t)0,
+        header.sampleRate,
+        header.bitsPerSample,
+        (i2s_channel_t)header.numChannels
+    );
+
+    // 查找 'data' 块
+    char chunkId[4];
+    uint32_t chunkSize;
+    int offset = sizeof(struct WavHeader);  // 初始偏移量为 WavHeader 的大小
+
+    // 循环读取每个块，直到找到 'data' 块
+    while (fread(chunkId, 4, 1, el->file) == 1) {
+        if (fread(&chunkSize, sizeof(chunkSize), 1, el->file) != 1) {
+            printf("read chunkSize faild.") ;
+            fclose(el->file);
+            el->file = NULL ;
+            return false ;
+        }
+
+        if (strncmp(chunkId, "data", 4) == 0) {
+            // 'data' 标记(4 字节) + 块大小(4 字节) = 8 字节
+            fseek(el->file, offset + 8, SEEK_SET);
+            return true ;
+        } else {
+            // 跳过非 'data' 块
+            fseek(el->file, chunkSize, SEEK_CUR);
+            offset += 8 + chunkSize;  // 8 字节为 chunkId 和 chunkSize 的大小
+        }
+    }
+
+    return false ;
 }